Moved graphics related code files into graphics directory
