+++ src/graphics/shader.cpp
@@ -1,4 +1,4 @@
-#include "shader.hpp"
+#include "graphics/shader.hpp"
 
 #include <iterator>
